+++ src/main/java/lotto/domain/strategy/LottoNumberStrategy.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.strategy;
 
 import java.util.List;
