+++ src/test/java/lotto/LottoMarketTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.LottoMarket;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
