+++ src/test/java/lotto/WinningNumbersTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.WinningNumbers;
 import org.junit.jupiter.api.Test;
 
 import java.util.Set;
